refactor(client): rendering
